PostgreSQL移行

+++ stationapi/src/use_case/dto/line.rs
@@ -11,9 +11,9 @@ impl From<Line> for GrpcLine {
             name_chinese: line.line_name_zh,
             name_korean: line.line_name_ko,
             color: line.line_color_c.unwrap_or_default(),
-            line_type: line.line_type.unwrap_or_default() as i32,
+            line_type: line.line_type.unwrap_or_default(),
             line_symbols: line.line_symbols.into_iter().map(|s| s.into()).collect(),
-            status: line.e_status as i32,
+            status: line.e_status,
             station: line.station.map(|s| Box::new(s.into())),
             company: line.company.map(|c| c.into()),
             train_type: line
